src/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
crispy-tailwind==0.2.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
idna==2.10
jsmin==2.2.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
